.prototyping/python/requirements.txt
scalekit-sdk-python>=1.0.0